+++ relayer/src/chain/cosmos.rs
@@ -81,6 +81,7 @@ const DEFAULT_GAS_FEE_AMOUNT: u64 = 1000;
 pub struct CosmosSdkChain {
     config: ChainConfig,
     rpc_client: HttpClient,
+    grpc_addr: Uri,
     rt: Arc<TokioRuntime>,
     keybase: KeyRing,
 }
@@ -90,13 +91,11 @@ impl CosmosSdkChain {
     pub fn unbonding_period(&self) -> Result<Duration, Error> {
         crate::time!("unbonding_period");
 
-        // TODO - generalize this
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
-                ibc_proto::cosmos::staking::v1beta1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::cosmos::staking::v1beta1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -327,26 +326,23 @@ impl Chain for CosmosSdkChain {
         let key_store = KeyRing::init(StoreBackend::Test, config.clone())
             .map_err(|e| Kind::KeyBase.context(e))?;
 
+        let grpc_addr =
+            Uri::from_str(&config.grpc_addr.to_string()).map_err(|e| Kind::Grpc.context(e))?;
+
         Ok(Self {
             rt,
             config,
             keybase: key_store,
             rpc_client,
+            grpc_addr,
         })
     }
 
-    // TODO use a simpler approach to create the light client
-    #[allow(clippy::type_complexity)]
-    fn init_light_client(
-        &self,
-    ) -> Result<(Box<dyn LightClient<Self>>, Option<thread::JoinHandle<()>>), Error> {
+    fn init_light_client(&self) -> Result<Box<dyn LightClient<Self>>, Error> {
         crate::time!("init_light_client");
 
-        let (lc, supervisor) = TMLightClient::from_config(&self.config, true)?;
-
-        let supervisor_thread = thread::spawn(move || supervisor.run().unwrap());
-
-        Ok((Box::new(lc), Some(supervisor_thread)))
+        let light_client = TMLightClient::from_config(&self.config)?;
+        Ok(Box::new(light_client))
     }
 
     fn init_event_monitor(
@@ -361,8 +357,11 @@ impl Chain for CosmosSdkChain {
     > {
         crate::time!("init_event_monitor");
 
-        let (mut event_monitor, event_receiver) =
-            EventMonitor::new(self.config.id.clone(), self.config.rpc_addr.clone(), rt)?;
+        let (mut event_monitor, event_receiver) = EventMonitor::new(
+            self.config.id.clone(),
+            self.config.websocket_addr.clone(),
+            rt,
+        )?;
 
         event_monitor.subscribe().unwrap();
         let monitor_thread = thread::spawn(move || event_monitor.run());
@@ -458,7 +457,7 @@ impl Chain for CosmosSdkChain {
 
         if status.sync_info.catching_up {
             fail!(
-                Kind::LightClientSupervisor(self.config.id.clone()),
+                Kind::LightClient(self.config.rpc_addr.to_string()),
                 "node at {} running chain {} not caught up",
                 self.config().rpc_addr,
                 self.config().id,
@@ -474,12 +473,11 @@ impl Chain for CosmosSdkChain {
     fn query_clients(&self, request: QueryClientStatesRequest) -> Result<Vec<ClientId>, Error> {
         crate::time!("query_chain_clients");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -525,13 +523,10 @@ impl Chain for CosmosSdkChain {
     ) -> Result<(Self::ClientState, MerkleProof), Error> {
         crate::time!("query_upgraded_client_state");
 
-        let grpc_address =
-            Uri::from_str(&self.config.grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
                 ibc_proto::cosmos::upgrade::v1beta1::query_client::QueryClient::connect(
-                    grpc_address,
+                    self.grpc_addr.clone(),
                 ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
@@ -576,13 +571,10 @@ impl Chain for CosmosSdkChain {
         let tm_height =
             Height::try_from(height.revision_height).map_err(|e| Kind::InvalidHeight.context(e))?;
 
-        let grpc_address =
-            Uri::from_str(&self.config.grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
                 ibc_proto::cosmos::upgrade::v1beta1::query_client::QueryClient::connect(
-                    grpc_address,
+                    self.grpc_addr.clone(),
                 ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
@@ -625,12 +617,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<AnyConsensusStateWithHeight>, Error> {
         crate::time!("query_chain_clients");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -654,11 +645,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<ConnectionId>, Error> {
         crate::time!("query_connections");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -687,11 +678,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<ConnectionId>, Error> {
         crate::time!("query_connections");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -730,11 +721,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<ChannelId>, Error> {
         crate::time!("query_connection_channels");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -760,11 +751,11 @@ impl Chain for CosmosSdkChain {
     fn query_channels(&self, request: QueryChannelsRequest) -> Result<Vec<ChannelId>, Error> {
         crate::time!("query_connections");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -809,11 +800,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<(Vec<PacketState>, ICSHeight), Error> {
         crate::time!("query_packet_commitments");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -842,11 +833,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<u64>, Error> {
         crate::time!("query_unreceived_packets");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -868,11 +859,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<(Vec<PacketState>, ICSHeight), Error> {
         crate::time!("query_packet_acknowledgements");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -901,11 +892,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<u64>, Error> {
         crate::time!("query_unreceived_acknowledgements");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -926,11 +917,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Sequence, Error> {
         crate::time!("query_next_sequence_receive");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -1425,11 +1416,11 @@ async fn broadcast_tx_commit(
 
 /// Uses the GRPC client to retrieve the account sequence
 async fn query_account(chain: &CosmosSdkChain, address: String) -> Result<BaseAccount, Error> {
-    let grpc_addr = Uri::from_str(&chain.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-    let mut client =
-        ibc_proto::cosmos::auth::v1beta1::query_client::QueryClient::connect(grpc_addr)
-            .await
-            .map_err(|e| Kind::Grpc.context(e))?;
+    let mut client = ibc_proto::cosmos::auth::v1beta1::query_client::QueryClient::connect(
+        chain.grpc_addr.clone(),
+    )
+    .await
+    .map_err(|e| Kind::Grpc.context(e))?;
 
     let request = tonic::Request::new(QueryAccountRequest { address });