+++ relayer/src/chain/handle.rs
@@ -64,12 +64,6 @@ pub enum ChainRequest {
         reply_to: ReplyTo<Vec<IbcEvent>>,
     },
 
-    GetMinimalSet {
-        from: Height,
-        to: Height,
-        reply_to: ReplyTo<Vec<AnyHeader>>,
-    },
-
     Signer {
         reply_to: ReplyTo<Signer>,
     },
@@ -90,6 +84,7 @@ pub enum ChainRequest {
     BuildHeader {
         trusted_height: Height,
         target_height: Height,
+        client_state: AnyClientState,
         reply_to: ReplyTo<AnyHeader>,
     },
 
@@ -99,11 +94,14 @@ pub enum ChainRequest {
     },
 
     BuildConsensusState {
-        height: Height,
+        trusted: Height,
+        target: Height,
+        client_state: AnyClientState,
         reply_to: ReplyTo<AnyConsensusState>,
     },
 
     BuildMisbehaviour {
+        client_state: AnyClientState,
         update_event: UpdateClient,
         latest_chain_height: Height,
         reply_to: ReplyTo<Option<AnyMisbehaviour>>,
@@ -241,8 +239,6 @@ pub trait ChainHandle: DynClone + Send + Sync + Debug {
     /// Send a transaction with `msgs` to chain.
     fn send_msgs(&self, proto_msgs: Vec<prost_types::Any>) -> Result<Vec<IbcEvent>, Error>;
 
-    fn get_minimal_set(&self, from: Height, to: Height) -> Result<Vec<AnyHeader>, Error>;
-
     fn get_signer(&self) -> Result<Signer, Error>;
 
     fn get_key(&self) -> Result<KeyEntry, Error>;
@@ -319,16 +315,23 @@ pub trait ChainHandle: DynClone + Send + Sync + Debug {
         &self,
         trusted_height: Height,
         target_height: Height,
+        client_state: AnyClientState,
     ) -> Result<AnyHeader, Error>;
 
     /// Constructs a client state at the given height
     fn build_client_state(&self, height: Height) -> Result<AnyClientState, Error>;
 
     /// Constructs a consensus state at the given height
-    fn build_consensus_state(&self, height: Height) -> Result<AnyConsensusState, Error>;
+    fn build_consensus_state(
+        &self,
+        trusted: Height,
+        target: Height,
+        client_state: AnyClientState,
+    ) -> Result<AnyConsensusState, Error>;
 
     fn build_misbehaviour(
         &self,
+        client_state: AnyClientState,
         update: UpdateClient,
         latest_chain_height: Height,
     ) -> Result<Option<AnyMisbehaviour>, Error>;