+++ relayer/src/chain/handle/prod.rs
@@ -86,10 +86,6 @@ impl ChainHandle for ProdChainHandle {
         })
     }
 
-    fn get_minimal_set(&self, from: Height, to: Height) -> Result<Vec<AnyHeader>, Error> {
-        self.send(|reply_to| ChainRequest::GetMinimalSet { from, to, reply_to })
-    }
-
     fn get_signer(&self) -> Result<Signer, Error> {
         self.send(|reply_to| ChainRequest::Signer { reply_to })
     }
@@ -229,10 +225,12 @@ impl ChainHandle for ProdChainHandle {
         &self,
         trusted_height: Height,
         target_height: Height,
+        client_state: AnyClientState,
     ) -> Result<AnyHeader, Error> {
         self.send(|reply_to| ChainRequest::BuildHeader {
             trusted_height,
             target_height,
+            client_state,
             reply_to,
         })
     }
@@ -241,16 +239,28 @@ impl ChainHandle for ProdChainHandle {
         self.send(|reply_to| ChainRequest::BuildClientState { height, reply_to })
     }
 
-    fn build_consensus_state(&self, height: Height) -> Result<AnyConsensusState, Error> {
-        self.send(|reply_to| ChainRequest::BuildConsensusState { height, reply_to })
+    fn build_consensus_state(
+        &self,
+        trusted: Height,
+        target: Height,
+        client_state: AnyClientState,
+    ) -> Result<AnyConsensusState, Error> {
+        self.send(|reply_to| ChainRequest::BuildConsensusState {
+            trusted,
+            target,
+            client_state,
+            reply_to,
+        })
     }
 
     fn build_misbehaviour(
         &self,
+        client_state: AnyClientState,
         update_event: UpdateClient,
         latest_chain_height: Height,
     ) -> Result<Option<AnyMisbehaviour>, Error> {
         self.send(|reply_to| ChainRequest::BuildMisbehaviour {
+            client_state,
             update_event,
             latest_chain_height,
             reply_to,