+++ relayer/src/foreign_client.rs
@@ -262,7 +262,7 @@ impl ForeignClient {
             .wrap_any();
 
         let consensus_state = self.src_chain
-            .build_consensus_state(latest_height)
+            .build_consensus_state(client_state.latest_height(),  latest_height, client_state.clone())
             .map_err(|e| ForeignClientError::ClientCreate(format!("failed while building client consensus state from src chain ({}) with error: {}", self.src_chain.id(), e)))?
             .wrap_any();
 
@@ -333,8 +333,8 @@ impl ForeignClient {
             thread::sleep(Duration::from_millis(100))
         }
 
-        // Get the latest trusted height from the client state on destination.
-        let trusted_height = self
+        // Get the latest client state on destination.
+        let client_state = self
             .dst_chain()
             .query_client_state(&self.id, Height::default())
             .map_err(|e| {
@@ -342,8 +342,9 @@ impl ForeignClient {
                     "failed querying client state on dst chain {} with error: {}",
                     self.id, e
                 ))
-            })?
-            .latest_height();
+            })?;
+
+        let trusted_height = client_state.latest_height();
 
         if trusted_height >= target_height {
             warn!(
@@ -355,7 +356,7 @@ impl ForeignClient {
 
         let header = self
             .src_chain()
-            .build_header(trusted_height, target_height)
+            .build_header(trusted_height, target_height, client_state)
             .map_err(|e| {
                 ForeignClientError::ClientUpdate(format!(
                     "failed building header with error: {}",
@@ -532,6 +533,17 @@ impl ForeignClient {
     ) -> Result<Option<AnyMisbehaviour>, ForeignClientError> {
         thread::sleep(Duration::from_millis(100));
 
+        // Get the latest client state on destination.
+        let client_state = self
+            .dst_chain()
+            .query_client_state(&self.id, Height::default())
+            .map_err(|e| {
+                ForeignClientError::Misbehaviour(format!(
+                    "failed querying client state on dst chain {} with error: {}",
+                    self.id, e
+                ))
+            })?;
+
         // get the list of consensus state heights in reverse order
         // Note: If chain does not prune consensus states then the last consensus state is
         // the one installed by the `CreateClient` which does not include a header.
@@ -561,7 +573,11 @@ impl ForeignClient {
 
             let misbehavior = self
                 .src_chain
-                .build_misbehaviour(update_event.clone(), latest_chain_height)
+                .build_misbehaviour(
+                    client_state.clone(),
+                    update_event.clone(),
+                    latest_chain_height,
+                )
                 .map_err(|e| {
                     ForeignClientError::Misbehaviour(format!("failed to build misbehaviour {}", e))
                 })?;